Commiting updates to fingerprint kernels following PR reviews.
